+++ DefenderAPI/internal/scripts/preimport.ps1
@@ -13,6 +13,8 @@ $moduleRoot = Split-Path (Split-Path $PSScriptRoot)
 # Load the strings used in messages
 "$moduleRoot\internal\scripts\strings.ps1"
 
+(Get-ChildItem "$moduleRoot\internal\classes\*.ps1" -ErrorAction Ignore).FullName
+
 # Load Tab Expansion
 (Get-ChildItem "$moduleRoot\internal\tepp\*.tepp.ps1" -ErrorAction Ignore).FullName